refactor: 만료일 도메인의 UI 종속성 제거
