+++ app/src/main/java/woowacourse/payments/ui/features/addcard/components/CardExpireDateField.kt
@@ -19,10 +19,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.R
-import woowacourse.payments.domain.card.values.ExpireDate.Companion.MAX_LENGTH_EXPIRE_DATE
 import woowacourse.payments.ui.components.AppTextField
 import woowacourse.payments.ui.features.addcard.ExpireDateUiState
 import woowacourse.payments.ui.mapper.messageResId
+import woowacourse.payments.ui.model.PaymentCardUiModel.Companion.MAX_EXPIRE_DATE_INPUT_LENGTH
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 import woowacourse.payments.ui.util.SeparatorVisualTransformation
 
@@ -41,12 +41,14 @@ fun CardExpireDateField(
     var isFocused by remember { mutableStateOf(false) }
     val showError =
         expireDateUiState is ExpireDateUiState.Invalid || expireDateUiState is ExpireDateUiState.Typing && !isFocused
-    val visualTransformation = remember { SeparatorVisualTransformation(EXPIRE_DATE_CHUNK_SIZE, EXPIRE_DATE_SEPARATOR) }
+    val visualTransformation =
+        remember { SeparatorVisualTransformation(EXPIRE_DATE_CHUNK_SIZE, EXPIRE_DATE_SEPARATOR) }
 
     AppTextField(
         value = value,
         onValueChange = { newValue ->
-            val filteredValue = newValue.filter { it in '0'..'9' }.take(MAX_LENGTH_EXPIRE_DATE)
+            val filteredValue =
+                newValue.filter { it in '0'..'9' }.take(MAX_EXPIRE_DATE_INPUT_LENGTH)
             onValueChange(filteredValue)
         },
         modifier =
