+++ app/src/main/java/woowacourse/payments/ui/mapper/CardCreationResult.kt
@@ -1,7 +1,7 @@
 package woowacourse.payments.ui.mapper
 
-import woowacourse.payments.domain.card.ExpireDateStatus
 import woowacourse.payments.domain.card.PaymentCard
+import woowacourse.payments.ui.model.ExpireDateStatus
 
 sealed class CardCreationResult {
     data class Success(
